Add support of hardlinking to api.file.copytree()
This CL adds a support of hardlinking of files to api.file.copytree()
API. With setting the argument to True, the API creates hardlinks of the
files, instead of copying files.

This CL just adds the functionally to the method, which is not called by
this CL yet, so that doesn't change any behavior.

Bug: 399504572
Test: Ran `/recipes.py test run` and `/unittests/run_test.py`
Test: Tested with led: https://ci.chromium.org/b/8721419464892830001
Change-Id: I86c67ca6934f3bcad52cb736327d77729cffcfac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/6309634
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>