[buildbucket] Check types in build_url()
Check the types of arguments to build_url(). It's relatively easy to
pass a build or build id into the first argument of this method, but
that's the host argument. Check the types so these kinds of errors are
caught in recipe testing instead of in production.

Change-Id: I91bcca63021376abe6901e48103645845e7a0d93
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/6650770
Auto-Submit: Rob Mohr <mohrr@google.com>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>