Revert "Add cert pinning to `url` recipe_module"
This reverts commit a71bdfef4b36fdad094e5635aa5d78fd356a44a1.

This certificate pinning didn’t achieve what it purported to: it doesn’t
actually pin to a specific intermediate, it actually pins to any root
listed in a certificate bundle .pem. Since leaf certificates may rotate
such that they may be signed by any root, and in practice the user of
this pinning has broken when leaves changed to be signed by different
roots, the pinning is both brittle and far broader than originally
expected.

I also believe that the specific operation of this pinning depends on
the details of the underlying TLS library. For example, in local
testing, I was able to pin to an intermediate without requiring full
chain verification to a trusted root (self-signed), but on the bots, the
verification would not be successful unless traceable to a trusted root,
even if the intermediate was present in the bundle.

Since the only known user of this pinning is being removed in
https://chrome-internal-review.googlesource.com/c/8586398, and the
pinning implementation doesn’t do what it was expected to, remove it to
avoid others taking a brittle and incompletely specified feature on as a
dependency.

README.recipes.md and test data were regenerated by running, from this
repository’s root directory:

./recipes.py test train --filter='url:*'

Bug: 444588116
Change-Id: Id7773f10ba8f9b5ae9d6367ff4638d6e4c29ab9b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/6944263
Reviewed-by: Avi Drissman <avi@chromium.org>
Auto-Submit: Mark Mentovai <mark@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Robbie Iannucci <iannucci@google.com>