[TurboCI] Add a post check graph assertion method.
This adds the ability to assert the state of the graph after the end
of the simulation. Failure looks like:

```
CHECK(FAIL):
  /.../recipes-py/recipe_engine/recipe_test_api.py:967 - post_check()
    `f(assert_, steps, *args, **kwargs)`
  /.../recipes-py/recipes/placeholder.py:423 - _assert_graph()
    `assert_((check_view.check.identifier.id == 'crazypants'))`
      check_view.check.identifier.id: 'bob'
  added /.../recipes-py/recipes/placeholder.py:430
    _assert_graph()
```

This also needs to implement the `types=['*']` pattern to be able
to pull the whole graph via the query_nodes API. This will be
available in the real service, but it will require an extra permission
there, as noted in the protos.

R=vadimsh

Bug: 449230833
Change-Id: I1ea65f05f08ca269c95bb06b98dce8aad353eaa3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/7049909
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Robbie Iannucci <iannucci@google.com>
Auto-Submit: Robbie Iannucci <iannucci@google.com>