[engine] Don't catch cancellations in except
Steps might be cancelled because they're in a self.m.time.timeout()
context, or because the overall build is being cancelled. The first case
is much like a failed step, and the caller likely wants to handle it as
a failed step. The latter case is not like a failed step, but more like
a KeyboardInterrupt. KeyboardInterrupt doesn't inherit from Exception,
so it's not caught by 'except Exception:' blocks. It inherits from
BaseException. This ensures that when a user hits Ctrl-C, very few
exception handlers are invoked to process the KeyboardInterrupt, and the
process simply exits.

Add a new exception type, CancelledBuild, that inherits directly from
BaseException. Add handling of this exception type in a couple places,
but mostly let it propagate untouched. This ensures that a cancelled
build won't be processed by an exception handler that doesn't understand
that the build has been cancelled.

Example failure: http://ci.chromium.org/b/8699098381363633537/infra

Bug: b/456536120
Change-Id: Iab371e7ada59b035efbddfe781a9e44017f7f463
Recipe-Manual-Change: build
Recipe-Manual-Change: chromiumos
Recipe-Manual-Change: fuchsia
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/7120818
Commit-Queue: Rob Mohr <mohrr@google.com>
Reviewed-by: Robbie Iannucci <iannucci@google.com>