Update turboci proto update script to exclude depot_tools protos
Currently, all non-internal, non-test protos are imported from the
turboci directory of the infra/turboci/proto repo. For protos that are
only used in downstream repos, this would make local development awkward
since the protos would need to be updated in the recipe_engine repo and
then the recipe dep would need to be overridden with the -O flag. This
excludes the depot_tools protos from being imported in favor of allowing
those to be imported separately into the depot_tools repo.

Bug: b:443496677
Change-Id: I3ca260001f284b3eb06659e86954711fb071d0a2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/7226090
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>