[TurboCI] Fix fake to have more internal coherency.
After the removal of OptionRef and ResultDataRef, there is no need to
have _db as a weird mix of Checks and Datum - it can just be a mapping
of Checks.

Fix this and other issues which logically fall out from this change.

Adds some more tests to ensure that the fact that the Check and its
data are stored in the same message doesn't bleed through the API
accidentally.

R=chanli, vadimsh

Bug: 467800005
Change-Id: I59cefb814d39fa1a68dc7fa44412d9e405742635
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/7277722
Commit-Queue: Robbie Iannucci <iannucci@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>